reworked the java doc comment added annotations and added remote checking of the user by ID

+++ src/main/java/web/dao/UserDAOImpl.java
@@ -23,7 +23,9 @@ public class UserDAOImpl implements UserDAO {
     @Override
     public void deleteUser(long id) {
         User user = entityManager.find(User.class, id);
-        entityManager.remove(user);
+        if(user == null){
+            entityManager.remove(user);
+        }
     }
 
     @Override
